+++ Assets/Scripts/World/WorldTime.cs
@@ -6,12 +6,12 @@ using System.Threading.Tasks;
 using Unity.Mathematics;
 
 static class WorldTime {
-	public static float GetTime(int ticks, float spinSpeed)
+	public static float GetTime(float ticks, float spinSpeed)
 	{
-		return math.modf((float)ticks * spinSpeed, out var i);
+		return math.modf(ticks * spinSpeed, out var i);
 	}
-	public static float GetDays(int ticks, float spinSpeed)
+	public static float GetDays(float ticks, float spinSpeed)
 	{
-		return (int)(ticks * spinSpeed);
+		return ticks * spinSpeed;
 	}
 }
